Fix(swagger): fix swagger doc not showing in UI
Descriptino:
During a test run for the swagger, I have noticed that there was
a problem with swagger configs which was causing the API doc to
not being loaded in swagger ui. After further investigation, I
have noticed two problems with swagger config. One was the
swagger's import in the main file, which was palced before ENV
files configuration. And the secound one was the apis option
value in swagger config file.

Change Log:
[Main file][Modified] - Relocate swagger's config import line
[Swagger config][Fixed] - Fix apis option value.